feat: 첫 화면 구현
